Use long for expires type (#24)

+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/BanPerson.kt
@@ -8,5 +8,5 @@ data class BanPerson(
     val ban: Boolean,
     val remove_data: Boolean? = null,
     val reason: String? = null,
-    val expires: Int? = null,
+    val expires: Long? = null,
 )
